the ios redux has all of the functions it needs to be broken down into parts
