--- ios/client/package.json
@@ -1,36 +1,0 @@
-{
-  "name": "Vocalize-ios-app",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {},
-  "author": "",
-  "license": "ISC",
-  "devDependencies": {
-    "autoprefixer-loader": "^2.1.0",
-    "babel-core": "^5.8.23",
-    "babel-eslint": "latest",
-    "babel-loader": "^5.3.2",
-    "css-loader": "^0.17.0",
-    "jsdom": "^6.5.1",
-    "react-hot-loader": "^1.3.0",
-    "style-loader": "^0.12.3",
-    "raw-loader": "latest",
-    "webpack": "^1.12.0",
-    "webpack-dev-server": "^1.10.1"
-  },
-  "dependencies": {
-    "axios": "latest",
-    "bootstrap": "^3.3.5",
-    "classnames": "^2.1.3",
-    "es6-promise": "^3.0.2",
-    "immutable": "^3.7.5",
-    "object-assign": "^4.0.1",
-    "react-native-audio": "^0.3.2",
-    "react-redux": "^3.1.0",
-    "redux": "^2.0.0",
-    "redux-react-router": "latest",
-    "redux-thunk": "latest",
-    "uuid": "^2.0.1"
-  }
-}
