--- ios/client/redux/middleware.js
@@ -1,1 +1,0 @@
-middleware.js
