+++ ios/package.json
@@ -6,9 +6,18 @@
     "start": "react-native start"
   },
   "dependencies": {
+    "lodash": "^3.10.1",
+    "moment": "^2.10.6",
+    "normalizr": "^1.3.1",
+    "react": "^0.14.0",
     "react-native": "^0.16.0",
     "react-native-audio": "^0.3.2",
     "react-native-button": "^1.3.1",
-    "react-native-vector-icons": "^1.0.2"
+    "react-native-vector-icons": "^1.0.2",
+    "react-pure-render": "^1.0.2",
+    "react-redux": "^3.1.0",
+    "redux": "^3.0.3",
+    "redux-logger": "^2.0.4",
+    "redux-thunk": "^1.0.0"
   }
 }
