+++ ios/src/constants/ActionTypes.js
@@ -1,6 +1,8 @@
 export const SET_STATE = 'SET_STATE';
 export const RECEIVE_NEW_WORD = 'RECEIVE_NEW_WORD';
+export const GET_SCORE = 'GET_SCORE'
 export const PLAY_WORD = 'PLAY_WORD';
+export const STOP_PLAYING = 'STOP_PLAYING';
 export const RECORD_USER_AUDIO = 'RECORD_USER_AUDIO';
 export const STOP_RECORDING_USER_AUDIO = 'STOP_RECORDING_USER_AUDIO';
 export const REQUEST_NEXT_WORD = 'REQUEST_NEXT_WORD';
