+++ ios/src/store/configure-store.js
@@ -1,0 +1,11 @@
+import {createStore, applyMiddleware} from 'redux';
+import thunkMiddleware from 'redux-thunk';
+import rootReducer from '../reducers/index';
+
+const createStoreWithMiddleware = applyMiddleware(thunkMiddleware)(createStore);
+
+export default function configureStore(initialState) {
+    const store = createStoreWithMiddleware(rootReducer, initialState);
+
+    return store;
+}
